HYDELL-UPD

+++ STM32/Inc/HY_MOD/motor/svgendq.h
@@ -11,7 +11,7 @@ typedef struct SVGENDQ	{
     float32_t   Va;     // Output: reference phase-a switching function		
     float32_t   Vb;     // Output: reference phase-b switching function 
     float32_t   Vc;     // Output: reference phase-c switching function
-    uint16_t    Sector; // Sector is treated as Q0 - independently with global Q
+    uint8_t     Sector; // Sector is treated as Q0 - independently with global Q
 } SVGENDQ;
 
 void SVGEN_run(SVGENDQ *svgq);
